tui lam them vài cho nhe

+++ NIIEPay/Dto/SavingRateDto.cs
@@ -1,5 +1,5 @@
 ﻿public class SavingRateDto
 {
     public int TermMonths { get; set; }
-    public double InterestRate { get; set; }
+    public decimal InterestRate { get; set; }
 }
